requirements.txt
tweepy
neo4j
textblob
colorama
python-dotenv
openai
requests-oauthlib
